+++ packages/mint-components/src/components/sqm-dropdown-field/useDropdownField.tsx
@@ -3,17 +3,12 @@ import {
   FormState,
   FORM_VALIDATION_CONTEXT,
 } from "../sqm-portal-register/useValidationState";
-import { NameFields } from "./sqm-name-fields";
 
-export function useNameFields(props: NameFields) {
+export function useDropdownField() {
   const validationState = useDomContext<FormState>(FORM_VALIDATION_CONTEXT);
   return {
     states: {
       validationErrors: validationState?.validationErrors,
-      content: {
-        lastNameLabel: props.lastNameLabel,
-        firstNameLabel: props.firstNameLabel,
-      },
     },
   };
 }