+++ packages/mint-components/src/components/sqm-stencilbook/sqm-stencilbook.tsx
@@ -47,6 +47,10 @@ import * as PortalFooter from "../sqm-portal-footer/PortalFooter.stories";
 import * as Hero from "../sqm-hero/Hero.stories";
 import * as ReferralIframe from "../sqm-referral-iframe/ReferralIframe.stories";
 import * as NameFields from "../sqm-name-fields/NameFields.stories";
+import * as CheckboxField from "../sqm-checkbox-field/CheckboxField.stories";
+import * as UseCheckboxField from "../sqm-checkbox-field/UseCheckboxField.stories";
+import * as DropdownField from "../sqm-dropdown-field/DropdownField.stories";
+import * as UseDropdownField from "../sqm-dropdown-field/UseDropdownField.stories";
 import * as RewardExchangeList from "../sqm-reward-exchange-list/RewardExchangeList.stories";
 import * as ProgramExplainer from "../sqm-program-explainer/ProgramExplainer.stories";
 import * as ProgramExplainerStep from "../sqm-program-explainer-step/ProgramExplainerStep.stories";
@@ -83,6 +87,8 @@ const stories = [
   UseBigStat,
   UseEditProfile,
   UseLeaderboard,
+  UseCheckboxField,
+  UseDropdownField,
   Router,
   PortalFrame,
   EditProfile,
@@ -114,6 +120,8 @@ const stories = [
   Hero,
   ReferralIframe,
   NameFields,
+  CheckboxField,
+  DropdownField,
   RewardExchangeList,
   UseRewardExchangeList,
   UseTaskCard,